реализовал PageController
